Added opportunity to read an entire rumd3 simulation in rumdpy format and example on how to calculate the rdf using rumdpy
in rumdpy/tools/load_output.py implemented load_rumd3 function which allows for reading the entire output contained in the rumd3 TrajectoryFiles (not restarts). This might be useful for testing of tools and comparing with old rumd3 versions

the example calc_rdf_from_rumd3.py uses load_output to read a rumd3 TrajectoryFiles and calculate the rdf. The rdf is calculated using the GPU