Complete implementation of nblist check for Lees-Edwards BC.
Remove last_box_shift and use extra paces in sim_box instead, so the interface for the nblist functinos becomes simpler again. Complete the dist_moved_sq_function for LE. However, somethings is wrong because the compte_plan seesm to fail and the checker ends up with a really slow configuration.