Add image variable to Lees-Edwards Simbox, let ConfSaver include simbox data with each configuration.
The device simbox arrayfor LeesEdwards  has now D+6 entries but on the host it's just D+2 and these are the ones that get written to the output. For an ordinary Simbox it's D entries of course. Whether to include simbox is set by the parameter include_simbox in the Simulation constructor.