Add function dist_moved_exceeds_limit_functino to Simbox to be used by nblist_check
The new function combines the previous Simbox function dist_moved_sq_function some code in update_strain change and some from nblist_check such that the essential comparison (basicall between distance moved, and the relevant skin) is done in the new function, which returns a boolean. The old code is still present and actually being used still. With the new version there seems to be an increase in neighbor-list rebuilds which I need to investigate.