Add flexibility to which quantities are computed inside the inner loop.
A dictionary, compute_flags, which is a class member, specifies which quantities to compute. compute_stresses is now part of this. pair_potential unfolds the different flags and relevant if statements handle the individual computations. The output phase has not been updated to handle this flexibility yet so code will not run if not all quantities are set (which is the current default).