Continue implementation of compute_flags
Simulation has its onw compute_flags which is logically distinct from that in Configuration but has the same defaults. A check is made that flags set True here should also be True in the configuration. A function to provide default flags is now used. Changes to the defaults can be given  to Configuration and Simulation constructors without having to give a complete compute_flags dictionary. All integrators now calculate KE and Fsq only if specified (unless needed for the algorithm as for example KE in NVT). All interactions get compute_flags from the Simulation. Tests and examples updated to use compute_flags instead of cmopute_stresses. Neighbor list gets compute_flags from the pair potential. Several fixed interactions are still missing the computation of u/w/stresses (though not all of these are necesarily meaningful for a given fixed interaction). Zeroing of stresses correctly implemented in both types of neighbor list. Evalutor uses the default compute_flags for now, but this likely needs to be considered.