Continue to implement flexible output, specifically in scalar_calculator
Flags Vol and Ptot added to compute_flags. Labels in scalar_calculator now match those in compute_flags in order and case, and the values of the compute_flags are used to determine what actually gets written to the output. extract_scalars has still hard-coded labels and column-IDs though, so there will be problems still when turning some of the quantities off.