Switch scalar labels in compute_flags to be identical with those in scalar_calculator
For example: 'u'->'U', 'k'->'K', 'lap'-> 'lapU'. Also, the function extract_scalars now uses the meta \
data ('scalar_names', previously called 'scalars_names') in file (or data structure) to get the labels\
 and columns indices. Since the name of scalar_names has been changed this means old data files would \
not be readable, but for now they are via an exception handling mechanism which will be removed in the\
 future. Computation of Fsq now included in SLLOD; and also in SLLOD a return statement now indented c\
onsistently with other integrators (no effect on correctness though).