Prepare further for reducing number of default scalars
Have removed almost all uses of make_configuration_cc in tests and examples, to allow specifying compute_flags when creating a Configuration. Not done in LJchain.py yet because there a number of particles is specified also. Relevant compute_flags specified now when you can create a Configuration explicitly. Another place where not possible yet is when getting a configuration from a file using configuration_from_hdf5. Put relevant if statements in bonds.py to check that U, W, lapU are to be computed.