added test_simbox.py and minor changes to Simbox.py
added tests/test_simbox.py : a test for the Simbox functions. Also tests a bit of Simbox_LeesEdwards but only partial for now.

in rumdpy/configuration/Simbox.py removed assert statements from docstrings because now they are in the test. Added short descriptions to each of the simbox functions. Commented out apply_PBC_dimension because not clear if of any use.