Replace single make_simbox_functions function with one for each function
This is preparation for making an abstract base class for Simbox. The required methods wil be get_dist_sq_dr_function(), etc. These now exist and are called elsewhere in the code to obtain a function object to be jitted, instead of storing them as attributes of the simbox class object.