Implement LEBC compatibility for nblist_linked_lists.
Some aspects are not quite optimal: now adds an extra cell at each end of the x-loop, where in principle only one extra is necessary. It thus requires at least 7 cells in the x-direction (to avoid double-counting). test_linked_lists_nblist.py has for now been given a hard-coded box_shift but that should be one of the parameters that hypothesis loops over. Some debuging code left in it for now, but commented out.