Generalize duplicate_molecule and duplicate_topologies to multiple types of molecule
The duplicate functions are still present for now, but there are new functions replicate_molecules and replicate_topologies. The user (for now) gives lists of topologies, positions, types and masses for the different molecule types, and a list specifying how many molecules of each type are desired. Also fix a bug in function topology.angles_from_bonds.