Finish implementing stress contribution from bonds (some bug fixes involved).
replicate_molecules takes a parameter compute_flags, to be passed to the newly created Configuration object. The molecules.py example now includes calculation of stress and compares the pressure obtained this way to that obtained using the virial. Warnings about stress not being cmoputed for angles and dihedrals. Fix bug in StressSaver where compute_flags was not set. Fix bug in Simulation.py regarding setting compute_flags based on those associated with the runtime_actions.